Creating files for workshop.

+++ src/prepareTrainingData.py
@@ -40,11 +40,11 @@ def writeH5(config):
     genome = pysam.FastaFile(config["genome"])
     outFile = h5py.File(config["output-h5"], "w")
     seqs = getSequences(regions, genome, outputLength, inputLength, jitter)
-    outFile.create_dataset("sequence", data=seqs)
+    outFile.create_dataset("sequence", data=seqs, dtype='i1')
     logging.debug("Sequence dataset created.")
     for i, head in enumerate(config["heads"]):
         headVals = getHead(regions, head["bigwig-files"], outputLength, jitter)
-        outFile.create_dataset("head_{0:d}".format(i), data=headVals)
+        outFile.create_dataset("head_{0:d}".format(i), data=headVals, dtype='f4')
         logging.debug("Added data for head {0:d}".format(i))
     outFile.close()
